feat, ci, build: library no longer header-only. Windows CIs now added. Project is now packagable
