+++ .gitignore
@@ -6,6 +6,7 @@
 _build
 build
 cmake-build-debug
+out
 
 # Xcode stuff
 *.xcworkspace
@@ -14,6 +15,3 @@ cmake-build-debug
 
 # Apple stuff.
 .DS_Store
-
-# Extras
-config.cfg
