+++ rconpp.pc.in
@@ -1,0 +1,11 @@
+prefix=@CMAKE_INSTALL_PREFIX@
+exec_prefix=${prefix}
+libdir=${prefix}/lib
+includedir=${prefix}/include
+
+Name: librconpp
+Description: A modern Source RCON library for C++
+Version: @PROJECT_VERSION@
+URL: https://github.com/Jaskowicz1/rconpp
+Libs: -L${libdir} -lrconpp
+Cflags: -I${includedir}
